+++ docs/.vitepress/config.ts
@@ -1,7 +1,7 @@
 /*
  * @Author: weisheng
  * @Date: 2023-07-27 10:26:09
- * @LastEditTime: 2023-08-22 19:02:54
+ * @LastEditTime: 2023-09-28 11:22:30
  * @LastEditors: weisheng
  * @Description: 
  * @FilePath: \wot-design-uni\docs\.vitepress\config.ts
@@ -35,14 +35,15 @@ export default defineConfig({
     },
     socialLinks: [
       { icon: 'github', link: 'https://github.com/Moonofweisheng/wot-design-uni' },
-      { icon: { svg: '<svg t="1692699544299" class="icon" viewBox="0 0 1024 1024" version="1.1" xmlns="http://www.w3.org/2000/svg" p-id="4184" width="200" height="200"><path d="M512 1024C230.4 1024 0 793.6 0 512S230.4 0 512 0s512 230.4 512 512-230.4 512-512 512z m259.2-569.6H480c-12.8 0-25.6 12.8-25.6 25.6v64c0 12.8 12.8 25.6 25.6 25.6h176c12.8 0 25.6 12.8 25.6 25.6v12.8c0 41.6-35.2 76.8-76.8 76.8h-240c-12.8 0-25.6-12.8-25.6-25.6V416c0-41.6 35.2-76.8 76.8-76.8h355.2c12.8 0 25.6-12.8 25.6-25.6v-64c0-12.8-12.8-25.6-25.6-25.6H416c-105.6 0-188.8 86.4-188.8 188.8V768c0 12.8 12.8 25.6 25.6 25.6h374.4c92.8 0 169.6-76.8 169.6-169.6v-144c0-12.8-12.8-25.6-25.6-25.6z" fill="#6D6D72" p-id="4185"></path></svg>' }, link: "https://gitee.com/fant-mini/wot-design-uni", ariaLabel: 'Gitee' }
+      { icon: { svg: '<svg t="1692699544299" class="icon" viewBox="0 0 1024 1024" version="1.1" xmlns="http://www.w3.org/2000/svg" p-id="4184" width="200" height="200"><path d="M512 1024C230.4 1024 0 793.6 0 512S230.4 0 512 0s512 230.4 512 512-230.4 512-512 512z m259.2-569.6H480c-12.8 0-25.6 12.8-25.6 25.6v64c0 12.8 12.8 25.6 25.6 25.6h176c12.8 0 25.6 12.8 25.6 25.6v12.8c0 41.6-35.2 76.8-76.8 76.8h-240c-12.8 0-25.6-12.8-25.6-25.6V416c0-41.6 35.2-76.8 76.8-76.8h355.2c12.8 0 25.6-12.8 25.6-25.6v-64c0-12.8-12.8-25.6-25.6-25.6H416c-105.6 0-188.8 86.4-188.8 188.8V768c0 12.8 12.8 25.6 25.6 25.6h374.4c92.8 0 169.6-76.8 169.6-169.6v-144c0-12.8-12.8-25.6-25.6-25.6z" fill="#6D6D72" p-id="4185"></path></svg>' }, link: "https://gitee.com/fant-mini/wot-design-uni", ariaLabel: 'Gitee' },
+      { icon: { svg: '<svg t="1694688365239" class="icon" viewBox="0 0 1024 1024" version="1.1" xmlns="http://www.w3.org/2000/svg" p-id="4048" width="200" height="200"><path d="M980.79827 694.105946c-21.144216-122.796973-109.844757-203.250162-109.844757-203.250162 12.647784-111.477622-33.792-131.26573-33.792-131.26573C827.392 14.668108 530.985514 20.67373 524.730811 20.839784 518.476108 20.67373 222.01427 14.668108 212.300108 359.590054c0 0-46.467459 19.788108-33.819676 131.26573 0 0-88.700541 80.453189-109.817081 203.250162 0 0-11.291676 207.484541 101.403676 25.40627 0 0 25.350919 69.161514 71.790703 131.26573 0 0-83.082378 28.256865-75.997405 101.625081 0 0-2.87827 81.836973 177.401081 76.218811 0 0 126.699243-9.852541 164.753297-63.515676l16.605405 0 0.276757 0 16.633081 0c38.026378 53.635459 164.725622 63.515676 164.725622 63.515676 180.224 5.618162 177.401081-76.218811 177.401081-76.218811 7.029622-73.368216-75.997405-101.625081-75.997405-101.625081 46.439784-62.104216 71.790703-131.26573 71.790703-131.26573C992.034595 901.590486 980.79827 694.105946 980.79827 694.105946z" p-id="4049" fill="#6D6D72"></path></svg>' }, link: "https://qm.qq.com/cgi-bin/qm/qr?k=O1Z3pal6StL39qHtABqR54Tb56igr90O&jump_from=webapi&authKey=MtVWfi/EQbT03wW7tKXv4bmyKYHBHtzI8VewlzSsOdxFjN0wbgNy17np9Z9yC4Z8", ariaLabel: 'QQ' }
     ],
     search: {
       provider: 'algolia',
       options: {
-        appId: '38TOUD5WW9',
-        apiKey: '06a0d60d393c7672523382347c8f9d55',
-        indexName: 'wot-design-uni',
+        appId: 'A74X2RFXSU',
+        apiKey: '6961856d63f5181bf71cb4fa3e4398d2',
+        indexName: 'wot-design-uni2',
       },
     },
 
@@ -59,11 +60,11 @@ export default defineConfig({
           },
           {
             text: '快速上手',
-            link: '/guide/quickUse',
+            link: '/guide/quick-use',
           },
           {
             text: '定制主题',
-            link: '/guide/customTheme',
+            link: '/guide/custom-theme',
           },
           {
             text: '字体',
@@ -75,7 +76,7 @@ export default defineConfig({
           },
           {
             text: '常见问题',
-            link: '/guide/commonProblems',
+            link: '/guide/common-problems',
           }
         ]
       },
@@ -93,7 +94,7 @@ export default defineConfig({
             link: "/component/calendar",
           }, {
             text: "反馈组件",
-            link: "/component/actionSheet",
+            link: "/component/action-sheet",
           }, {
             text: "数据展示",
             link: "/component/badge",
@@ -122,11 +123,11 @@ export default defineConfig({
         },
         {
           text: '快速上手',
-          link: '/guide/quickUse',
+          link: '/guide/quick-use',
         },
         {
           text: '定制主题',
-          link: '/guide/customTheme',
+          link: '/guide/custom-theme',
         },
         {
           text: '字体',
@@ -138,7 +139,7 @@ export default defineConfig({
         },
         {
           text: '常见问题',
-          link: '/guide/commonProblems',
+          link: '/guide/common-problems',
         }
       ],
       '/reward/': [
@@ -151,7 +152,7 @@ export default defineConfig({
           link: '/reward/donor',
         },
       ],
-
+    
       '/component/': [
         {
           text: '基础',
@@ -166,7 +167,7 @@ export default defineConfig({
               link: "/component/layout",
               text: "Layout 布局"
             }, {
-              link: "/component/configProvider",
+              link: "/component/config-provider",
               text: "ConfigProvider 全局配置"
             }, {
               link: "/component/popup",
@@ -181,7 +182,7 @@ export default defineConfig({
           ]
         },
         {
-
+    
           text: "导航",
           items: [{
             link: "/component/pagination",
@@ -192,27 +193,30 @@ export default defineConfig({
           }, {
             link: "/component/tabs",
             text: "Tabs 标签页"
+          }, {
+            link: "/component/segmented",
+            text: "Segmented 分段器"
           }]
         }, {
-
+    
           text: "数据输入",
           items: [{
             link: "/component/calendar",
             text: "Calendar 日历选择器"
           }, {
-            link: "/component/calendarView",
+            link: "/component/calendar-view",
             text: "CalendarView 日历面板"
           }, {
             link: "/component/checkbox",
             text: "Checkbox 复选框"
           }, {
-            link: "/component/colPicker",
+            link: "/component/col-picker",
             text: "ColPicker 多列选择器"
           }, {
-            link: "/component/datetimePicker",
+            link: "/component/datetime-picker",
             text: "DatetimePicker 时间选择器"
           }, {
-            link: "/component/datetimePickerView",
+            link: "/component/datetime-picker-view",
             text: "DatetimePickerView 时间选择器视图"
           }, {
             link: "/component/form",
@@ -221,13 +225,13 @@ export default defineConfig({
             link: "/component/input",
             text: "Input 输入框"
           }, {
-            link: "/component/inputNumber",
+            link: "/component/input-number",
             text: "InputNumber 计数器"
           }, {
             link: "/component/picker",
             text: "Picker 选择器"
           }, {
-            link: "/component/pickerView",
+            link: "/component/picker-view",
             text: "PickerView 选择器视图"
           }, {
             link: "/component/radio",
@@ -239,7 +243,7 @@ export default defineConfig({
             link: "/component/search",
             text: "Search 搜索框"
           }, {
-            link: "/component/selectPicker",
+            link: "/component/select-picker",
             text: "SelectPicker 单复选选择器"
           }, {
             link: "/component/slider",
@@ -254,31 +258,34 @@ export default defineConfig({
         }, {
           text: "反馈",
           items: [{
-            link: "/component/actionSheet",
+            link: "/component/action-sheet",
             text: "ActionSheet 动作面板"
           }, {
-            link: "/component/dropMenu",
+            link: "/component/drop-menu",
             text: "DropMenu 下拉菜单"
           }, {
             link: "/component/loading",
             text: "Loading 加载"
           }, {
-            link: "/component/messageBox",
+            link: "/component/message-box",
             text: "MessageBox 弹框"
           }, {
-            link: "/component/noticeBar",
+            link: "/component/notice-bar",
             text: "NoticeBar 通知栏"
           }, {
             link: "/component/progress",
             text: "Progress 进度条"
           }, {
-            link: "/component/sortButton",
+            link: "/component/circle",
+            text: "Circle 环形进度条"
+          }, {
+            link: "/component/sort-button",
             text: "SortButton 排序按钮"
           }, {
-            link: "/component/statusTip",
+            link: "/component/status-tip",
             text: "StatusTip 缺省提示"
           }, {
-            link: "/component/swipeAction",
+            link: "/component/swipe-action",
             text: "SwipeAction 滑动操作"
           }, {
             link: "/component/toast",
@@ -291,7 +298,7 @@ export default defineConfig({
             text: "Tooltip 文字提示"
           }]
         }, {
-
+    
           text: "数据展示",
           items: [{
             link: "/component/badge",
@@ -315,7 +322,7 @@ export default defineConfig({
             link: "/component/img",
             text: "Img 图片"
           }, {
-            link: "/component/imgCropper",
+            link: "/component/img-cropper",
             text: "ImgCropper 图片裁剪"
           }, {
             link: "/component/grid",
@@ -335,10 +342,17 @@ export default defineConfig({
           }, {
             link: "/component/tag",
             text: "Tag 标签"
+          }, {
+            link: "/component/watermark",
+            text: "Watermark 水印"
+          }, {
+            link: "/component/swiper",
+            text: "Swiper 轮播图"
           }]
         }
       ]
     }
+    
   },
 
 })