+++ src/uni_modules/wot-design-uni/components/common/util.ts
@@ -16,22 +16,21 @@ function s4() {
 }
 
 /**
- * @description 将数字自动转换为带有 'px' 单位的字符串
- * @param {number | string} num - 要转换的数字
- * @return {string} 转换后的字符串，带有 'px' 单位
+ * @description 对num自动填充px
+ * @param {Number} num
+ * @return {string} num+px
  */
-export function addUnit(num: number | string): string {
-  const convertedNumber = typeof num === 'number' ? num : parseFloat(num)
-  return isNaN(convertedNumber) ? String(num) : `${convertedNumber}px`
+export function addUnit(num: number | string) {
+  return Number.isNaN(Number(num)) ? num : `${num}px`
 }
 
 /**
  * @description 判断target是否对象
- * @param obj
+ * @param value
  * @return {boolean}
  */
-export function isObj(obj: any): boolean {
-  return Object.prototype.toString.call(obj) === '[object Object]' || typeof obj === 'object'
+export function isObj(value: any): value is object {
+  return Object.prototype.toString.call(value) === '[object Object]' || typeof value === 'object'
 }
 
 /**
@@ -342,6 +341,16 @@ export function isBoolean(value: any): value is boolean {
 }
 
 /**
+ * 是否为base64图片
+ * @param {string} url
+ * @return
+ */
+export function isBase64Image(url: string) {
+  // 使用正则表达式检查URL是否以"data:image"开头，这是Base64图片的常见前缀
+  return /^data:image\/(png|jpg|jpeg|gif|bmp);base64,/.test(url)
+}
+
+/**
  * 将外部传入的样式格式化为可读的 CSS 样式。
  * @param {object | object[]} styles 外部传入的样式对象或数组
  * @returns {string} 格式化后的 CSS 样式字符串
@@ -465,4 +474,23 @@ export function deepMerge<T extends Record<string, any>>(target: T, source: Reco
   }
 
   return target
+}
+
+/**
+ * 构建带参数的URL
+ * @param baseUrl 基础URL
+ * @param params 参数对象，键值对表示要添加到URL的参数
+ * @returns 返回带有参数的URL
+ */
+export function buildUrlWithParams(baseUrl: string, params: Record<string, string>) {
+  // 将参数对象转换为查询字符串
+  const queryString = Object.entries(params)
+    .map(([key, value]) => `${key}=${encodeURIComponent(value)}`)
+    .join('&')
+
+  // 检查基础URL是否已包含查询字符串，并选择适当的分隔符
+  const separator = baseUrl.includes('?') ? '&' : '?'
+
+  // 返回带有参数的URL
+  return `${baseUrl}${separator}${queryString}`
 }