+++ src/uni_modules/wot-design-uni/components/wd-grid-item/wd-grid-item.vue
@@ -29,7 +29,7 @@ export default {
 </script>
 
 <script lang="ts" setup>
-import { getCurrentInstance, inject, onBeforeMount, onMounted, ref, nextTick, computed, Ref, watch } from 'vue'
+import { getCurrentInstance, inject, onBeforeMount, onMounted, ref, type Ref, watch } from 'vue'
 
 type BadgeType = 'primary' | 'success' | 'warning' | 'danger' | 'info'
 type LinkType = 'navigateTo' | 'switchTab' | 'reLaunch' | 'redirectTo'
@@ -126,13 +126,37 @@ function init() {
   style.value = `width: ${width}; ${squareStyle || gutterStyle}`
 }
 
-function click(event) {
+function click() {
   if (!parent.clickable) return
   const { url, linkType } = props
-  if (url) {
-    ;(uni as any)[linkType]({ url })
-  }
   emit('itemclick')
+  if (url) {
+    switch (linkType) {
+      case 'navigateTo':
+        uni.navigateTo({
+          url
+        })
+        break
+      case 'reLaunch':
+        uni.reLaunch({
+          url
+        })
+        break
+      case 'redirectTo':
+        uni.redirectTo({
+          url
+        })
+        break
+      case 'switchTab':
+        uni.switchTab({
+          url
+        })
+        break
+      default:
+        console.warn(`[wot-design] warning(wd-grid-item): linkType can not be ${linkType}`)
+        break
+    }
+  }
 }
 /**
  * 设置样式