+++ src/uni_modules/wot-design-uni/components/wd-message-box/index.ts
@@ -1,14 +1,14 @@
 /*
  * @Author: weisheng
  * @Date: 2022-12-14 17:33:21
- * @LastEditTime: 2023-06-19 12:52:04
+ * @LastEditTime: 2023-09-07 00:29:51
  * @LastEditors: weisheng
  * @Description:
  * @FilePath: \wot-design-uni\src\uni_modules\wot-design-uni\components\wd-message-box\index.ts
  * 记得注释
  */
-import { InjectionKey, Ref, provide, ref } from 'vue'
-import { Message, MessageOptions, MessageResult, MessageType } from './types'
+import { provide, ref } from 'vue'
+import type { Message, MessageOptions, MessageResult, MessageType } from './types'
 import { deepMerge } from '../common/util'
 
 /**
@@ -16,8 +16,7 @@ import { deepMerge } from '../common/util'
  *
  * @internal
  */
-export const messageDefaultKey = Symbol('__MESSAGE__') as InjectionKey<Ref<boolean>>
-export const messageDefaultOptionKey = Symbol('__MESSAGE_OPTION__') as InjectionKey<Ref<MessageOptions>>
+export const messageDefaultOptionKey = '__MESSAGE_OPTION__'
 
 // 默认模板
 export const defaultOptions: MessageOptions = {
@@ -41,13 +40,19 @@ export const defaultOptions: MessageOptions = {
 
 export function useMessage(selector: string = ''): Message {
   const messageOption = ref<MessageOptions>(defaultOptions) // Message选项
-  const messageOptionKey = selector ? '__MESSAGE_OPTION__' + selector : messageDefaultOptionKey
+  const messageOptionKey = selector ? messageDefaultOptionKey + selector : messageDefaultOptionKey
   provide(messageOptionKey, messageOption)
 
   const createMethod = (type: MessageType) => {
     // 优先级：options->toastOptions->defaultOptions
     return (options: MessageOptions | string) => {
-      return show(deepMerge({ type: type }, typeof options === 'string' ? { title: options } : options) as MessageOptions)
+      const messageOptions = deepMerge({ type: type }, typeof options === 'string' ? { title: options } : options) as MessageOptions
+      if (messageOptions.type === 'confirm' || messageOptions.type === 'prompt') {
+        messageOptions.showCancelButton = true
+      } else {
+        messageOptions.showCancelButton = false
+      }
+      return show(messageOptions)
     }
   }